+++ Libplanet/Action/GasMeter.cs
@@ -2,10 +2,10 @@ namespace Libplanet.Action
 {
     internal class GasMeter : IGasMeter
     {
-        public GasMeter(long gasLimit)
+        public GasMeter(long gasLimit, long gasUsed = 0)
         {
             SetGasLimit(gasLimit);
-            GasUsed = 0;
+            GasUsed = gasUsed;
         }
 
         public long GasAvailable => GasLimit - GasUsed;
@@ -33,6 +33,7 @@ namespace Libplanet.Action
 
             if (newGasUsed > GasLimit)
             {
+                GasUsed = GasLimit;
                 throw new GasLimitExceededException(GasLimit, newGasUsed);
             }