+++ CHANGES.md
@@ -13,11 +13,18 @@ To be released.
  -  Added `IAccountStateDelta.TotalUpdatedFungibleAssets` interface property.
     [[#3208]]
 
+ -  Added `ITransaction.MaxGasPrice` property.  [[#3201]]
+ -  Added `ITransaction.GasLimit` property.  [[#3201]]
+ -  `Currency.Serialize()`'s behavioral is changed.
+    the serialize field is replaced. [[#3201]]
+
 ### Backward-incompatible network protocol changes
 
 ### Backward-incompatible storage format changes
 
 ### Added APIs
+
+ -  Added `FungibleAssetValue.Serialize()` method.  [[#3201]]
 
 ### Behavioral changes